Move ownership of biometrics to key-management
