+++ .github/CODEOWNERS
@@ -100,11 +100,19 @@ apps/desktop/src/services/native-message-handler.service.ts @bitwarden/team-auto
 ## Component Library ##
 .storybook @bitwarden/team-design-system
 libs/components @bitwarden/team-design-system
-apps/browser/src/platform/popup/layout @bitwarden/team-design-system 
-apps/web/src/app/layouts @bitwarden/team-design-system 
+apps/browser/src/platform/popup/layout @bitwarden/team-design-system
+apps/web/src/app/layouts @bitwarden/team-design-system
 
 ## Desktop native module ##
 apps/desktop/desktop_native @bitwarden/team-platform-dev
+
+## Key management team files ##
+apps/desktop/destkop_native/src/biometric/ @bitwarden/team-key-management-dev
+apps/desktop/src/key-management @bitwarden/team-key-management-dev
+apps/web/src/key-management @bitwarden/team-key-management-dev
+apps/browser/src/key-management @bitwarden/team-key-management-dev
+apps/cli/src/key-management @bitwarden/team-key-management-dev
+
 
 ## DevOps team files ##
 /.github/workflows @bitwarden/dept-devops
