+++ apps/desktop/src/key-management/biometrics/biometrics.service.ts
@@ -1,10 +1,10 @@
+import { BiometricStateService } from "@bitwarden/common/key-management/biometrics/biometric-state.service";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
 import { LogService } from "@bitwarden/common/platform/abstractions/log.service";
 import { MessagingService } from "@bitwarden/common/platform/abstractions/messaging.service";
-import { BiometricStateService } from "@bitwarden/common/platform/biometrics/biometric-state.service";
 import { UserId } from "@bitwarden/common/types/guid";
 
-import { WindowMain } from "../../../main/window.main";
+import { WindowMain } from "../../main/window.main";
 
 import { DesktopBiometricsService, OsBiometricService } from "./desktop.biometrics.service";
