+++ apps/desktop/src/key-management/biometrics/electron-biometrics.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from "@angular/core";
 
-import { BiometricsService } from "@bitwarden/common/platform/biometrics/biometric.service";
+import { BiometricsService } from "@bitwarden/common/key-management/biometrics/biometric.service";
 
 /**
  * This service implement the base biometrics service to provide desktop specific functions,
