+++ libs/common/src/platform/services/system.service.ts
@@ -7,11 +7,11 @@ import { AuthService } from "../../auth/abstractions/auth.service";
 import { AuthenticationStatus } from "../../auth/enums/authentication-status";
 import { AutofillSettingsServiceAbstraction } from "../../autofill/services/autofill-settings.service";
 import { VaultTimeoutAction } from "../../enums/vault-timeout-action.enum";
+import { BiometricStateService } from "../../key-management/biometrics/biometric-state.service";
 import { UserId } from "../../types/guid";
 import { MessagingService } from "../abstractions/messaging.service";
 import { PlatformUtilsService } from "../abstractions/platform-utils.service";
 import { SystemService as SystemServiceAbstraction } from "../abstractions/system.service";
-import { BiometricStateService } from "../biometrics/biometric-state.service";
 import { Utils } from "../misc/utils";
 import { ScheduledTaskNames } from "../scheduling/scheduled-task-name.enum";
 import { TaskSchedulerService } from "../scheduling/task-scheduler.service";
